Move utility methods to VitDeck.Utilities
