+++ VitDeck/Assets/VitDeck/Utilities/AssetUtility.cs
@@ -1,3 +1,4 @@
+using System.Collections.Generic;
 using System.IO;
 using UnityEditor;
 
@@ -34,6 +35,28 @@ namespace VitDeck.Utilities
                     }
                 }
                 return _imageFolderPath;
+            }
+        }
+
+        /// <summary>
+        /// 指定されたファイル/ディレクトリ及び、その子のアセットを列挙する。
+        /// </summary>
+        /// <param name="path">アセットもしくはディレクトリのパス</param>
+        /// <returns>アセットの列挙</returns>
+        public static IEnumerable<UnityEngine.Object> EnumerateAssets(string path)
+        {
+            foreach (var assetPath in IOUtility.EnumerateFileSystemEntries(path))
+            {
+                if (assetPath.EndsWith(".meta"))
+                {
+                    continue;
+                }
+                var asset = AssetDatabase.LoadAssetAtPath<UnityEngine.Object>(assetPath);
+                if (asset == null)
+                {
+                    continue;
+                }
+                yield return asset;
             }
         }
     }
