+++ VitDeck/Assets/VitDeck/Utilities/IOUtility.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 41437640cbfab4a40b9a3cd78700bd06
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
